pyproject.toml
[tool.poetry]
name = "scheduling-api"
version = "0.1.0"
description = "Multi-tenant scheduling API for businesses"
authors = ["Jacqueline Diniz <jacquelinediniz.tech@gmail.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = ">=0.115.0"
uvicorn = {extras = ["standard"], version = ">=0.34.0"}
pydantic = {extras = ["email"], version = "^2.13.3"}
pydantic-settings = "^2.14.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.49"}
asyncpg = "^0.31.0"
alembic = "^1.18.4"
python-dotenv = "^1.2.2"
greenlet = "^3.5.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.27"
bcrypt = "^5.0.0"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.4.0"
black = ">=24.0.0"
pytest = ">=8.0.0"
pytest-asyncio = ">=0.23.0"
httpx = ">=0.27.0"
pre-commit = ">=3.7.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.black]
line-length = 88